requirements.txt
altgraph==0.17
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==7.1.2
data==0.4
decorator==5.1.1
Django==4.0.2
django-crispy-forms==1.14.0
django-reset-migrations==0.4.0
fake-useragent==0.1.11
Flask==1.1.2
funcsigs==1.0.2
future==0.18.2
gunicorn==20.1.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
numpy==1.18.3
pefile==2019.4.18
Pillow==9.0.1
PyMySQL==0.9.3
pywin32-ctypes==0.2.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
style==1.1.0
tzdata==2021.5
update==0.0.1
urllib3==1.26.8
vsearch==1.0
Werkzeug==1.0.1
